backend/requirements.txt
# This file is now in the backend directory for Docker builds

alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.9.3
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.3.2
click==8.1.8
cycler==0.11.0
distlib==0.3.4
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.4.2
fonttools==4.29.1
future==0.18.3
git-filter-repo==2.47.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
imutils==0.5.4
jiter==0.10.0
joblib==1.4.2
kiwisolver==1.3.2
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.5.1
MouseInfo==0.1.3
numpy==1.26.4
ollama==0.5.1
ollama-python==0.1.2
openai==1.84.0
opencv-python==4.5.2.54
openpyxl==3.0.9
outcome==1.3.0.post0
packaging==21.3
pandas==1.3.4
panorama==0.5.0
pillow==11.2.1
platformdirs==2.4.1
PyAutoGUI==0.9.54
pydantic==2.11.5
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.0.7
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-resize-image==1.1.20
pytweening==1.2.0
pytz==2021.3
PyYAML==5.3.1
requests==2.31.0
responses==0.24.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
rubicon-objc==0.5.1
scikit-build==0.17.6
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.33.0
six==1.16.0
sklearn==0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.35.0
validate-email==1.3
virtualenv==20.13.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
